+++ emuhelper/cp2k/opt.py
@@ -25,34 +25,52 @@ class opt_run:
         self.motion = cp2k_motion()
         
         self.run_type = "GEO_OPT" # CELL_OPT
-        self.set_run_type("GEO_OPT")
 
         cutoff = 60
         rel_cutoff = 30
 
         
-    def gen_input(self, directory="tmp-cp2k-opt", inpname="geometric-optimization.inp"):
+    def geo_opt(self, directory="tmp-cp2k-geo-opt", inpname="geo-opt.inp", output="geo-opt.out", mpi="", runopt="gen"):
         """
         directory: a place for all the generated files
         """
-        if os.path.exists(directory):
-            shutil.rmtree(directory)
-        os.mkdir(directory)
-        shutil.copyfile(self.force_eval.subsys.xyz.file, os.path.join(directory, self.force_eval.subsys.xyz.file))
+        if runopt == "gen" or runopt == "genrun":
+            if os.path.exists(directory):
+                shutil.rmtree(directory)
+            os.mkdir(directory)
+            shutil.copyfile(self.force_eval.subsys.xyz.file, os.path.join(directory, self.force_eval.subsys.xyz.file))
 
-        self.glob.to_input(os.path.join(directory, inpname))
-        self.force_eval.to_input(os.path.join(directory, inpname))
-        self.motion.to_input(os.path.join(directory, inpname))
+            self.set_geo_opt()
+            self.glob.to_input(os.path.join(directory, inpname))
+            self.force_eval.to_input(os.path.join(directory, inpname))
+            self.motion.to_input(os.path.join(directory, inpname))
+        
+        if runopt == "run" or runopt == "genrun":
+            os.chdir(directory)
+            os.system("%s cp2k.psmp -in %s | tee %s" % (mpi, inpname, output))
+            os.chdir("../")
     
-    def run(self, directory="tmp-cp2k-opt", inpname="geometric-optimization.inp", output="geometric-optimization.out"):
+    def cell_opt(self, directory="tmp-cp2k-cell-opt", inpname="cell-opt.inp", output="cell-opt.out", mpi="", runopt="gen"):
         """
         directory: a place for all the generated files
         """
-        os.chdir(directory)
-        os.system("cp2k.psmp -in %s | tee %s" % (inpname, output))
-        os.chdir("../")    
+        if runopt == "gen" or runopt == "genrun":
+            if os.path.exists(directory):
+                shutil.rmtree(directory)
+            os.mkdir(directory)
+            shutil.copyfile(self.force_eval.subsys.xyz.file, os.path.join(directory, self.force_eval.subsys.xyz.file))
+            
+            self.set_cell_opt()
+            self.glob.to_input(os.path.join(directory, inpname))
+            self.force_eval.to_input(os.path.join(directory, inpname))
+            self.motion.to_input(os.path.join(directory, inpname))
+        
+        if runopt == "run" or runopt == "genrun":
+            os.chdir(directory)
+            os.system("%s cp2k.psmp -in %s | tee %s" % (mpi, inpname, output))
+            os.chdir("../")
 
-    def analysis(self, directory="tmp-cp2k-opt", output="geometric-optimization.out"):
+    def analysis(self, directory="tmp-cp2k-geo-opt", output="geo-opt.out"):
         # analyse the result
         os.chdir(directory)
         os.system("cat %s | grep 'ENERGY| Total FORCE_EVAL' > energy-per-ion-step.data" % (output))
@@ -67,20 +85,17 @@ class opt_run:
 
         os.chdir("../")
     
-    def set_run_type(self, run_type="GEO_OPT"):
-        # run_type can only be one of "GEO_OPT" and "CELL_OPT"
+    def set_geo_opt(self):
+        self.run_type = "GEO_OPT"
+        self.glob.params["RUN_TYPE"] = "GEO_OPT"
+        self.motion.set_type("GEO_OPT")
+    
+    def set_cell_opt(self):
         # Note:
         #   if you are doing CELL_OPT run, you must also enable
         #   "STRESS_TENSOR" in FORCE_EVAL%STRESS_TENSOR
-        if run_type != "GEO_OPT" and run_type != "CELL_OPT":
-            print("==========================================\n")
-            print("           WARNING    !!!!!!!!!\n")
-            print("==========================================\n")
-            print("cp2k.opt.opt_run can only conduct 'GEO_OPT'\n")
-            print("and 'CELL_OPT'\n")
-            sys.exit(1)
-        self.run_type = run_type 
-        self.glob.params["RUN_TYPE"] = run_type
-        self.motion.set_type(run_type)
-        if run_type == "CELL_OPT" and self.force_eval.params["STRESS_TENSOR"] is None:
+        self.run_type = "CELL_OPT" 
+        self.glob.params["RUN_TYPE"] = "CELL_OPT"
+        self.motion.set_type("CELL_OPT")
+        if self.force_eval.params["STRESS_TENSOR"] is None:
             self.force_eval.params["STRESS_TENSOR"] = "ANALYTICAL" # NUMERICAL
